refactor: migrate to nestjs
