+++ package.json
@@ -1,50 +1,78 @@
 {
-  "name": "did-core",
+  "name": "nest-typescript-starter",
+  "private": true,
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "Nest TypeScript starter repository",
+  "license": "MIT",
   "scripts": {
-    "test": "jest --detectOpenHandles --forceExit",
-    "build": "tsc",
-    "start": "tsc; node dist/index.js"
+    "prebuild": "rimraf dist",
+    "build": "nest build",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+    "start": "nest start",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
+    "start:prod": "node dist/main",
+    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+    "test:e2e": "jest --config ./test/jest-e2e.json"
   },
-  "jest": {
-    "transform": {
-      "^.+\\.ts$": "ts-jest"
-    },
-    "testRegex": "\\.test\\.ts$",
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "json"
-    ],
-    "globals": {
-      "ts-jest": {
-        "diagnostics": true
-      }
-    }
-  },
-  "author": "",
-  "license": "ISC",
   "dependencies": {
+    "@nestjs/common": "^8.1.1",
+    "@nestjs/core": "^8.1.1",
+    "@nestjs/platform-express": "^8.1.1",
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^7.4.0",
     "@ethersproject/providers": "^5.6.6",
     "cors": "^2.8.5",
     "did-jwt-vc": "^2.1.9",
     "ethr-did": "^2.2.2",
-    "express": "^4.18.1",
     "http-status-codes": "^2.2.0",
     "morgan": "^1.10.0"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.12",
+    "@nestjs/cli": "^8.1.3",
+    "@nestjs/schematics": "^8.0.4",
+    "@nestjs/testing": "^8.1.1",
     "@types/express": "^4.17.13",
-    "@types/jest": "^27.5.1",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.11.1",
+    "@types/supertest": "^2.0.11",
+    "@types/cors": "^2.8.12",
     "@types/morgan": "^1.9.3",
-    "@types/node": "^17.0.35",
-    "jest": "^28.1.0",
-    "supertest": "^6.2.3",
-    "ts-jest": "^28.0.2",
-    "typescript": "^4.6.4"
+    "@typescript-eslint/eslint-plugin": "^4.29.2",
+    "@typescript-eslint/parser": "^4.29.2",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.1",
+    "jest": "^27.3.0",
+    "prettier": "^2.4.1",
+    "source-map-support": "^0.5.20",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.0.7",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.3.0",
+    "tsconfig-paths": "^3.11.0",
+    "typescript": "^4.4.4"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "ts"
+    ],
+    "rootDir": "src",
+    "testRegex": ".*\\.spec\\.ts$",
+    "transform": {
+      "^.+\\.(t|j)s$": "ts-jest"
+    },
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
+    ],
+    "coverageDirectory": "../coverage",
+    "testEnvironment": "node"
   }
 }
